requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
reportlab
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
